fix(rta): bound the ECP readings so failure capture cannot hang
`captureFailureState()` bounded its two ODC reads at 5 s but left both ECP
readings unbounded, and `.catch()` handles a rejection, not a hang. Since they
sit in the same `Promise.all`, one unsettled promise withholds the entire
capture — and this runs on EVERY wait timeout, which is exactly when the device
is least likely to be healthy. The result would be the worst shape available: a
suite that hangs instead of reporting the timeout it had already detected.

The in-code claim that ECP "options carry no timeout" is correct and was
verified, but it did not follow that the call was bounded. Read out of the
installed client on 2026-09-08: `HttpRequestOptions` exposes only `retryCount`
(`RokuDevice.d.ts:6-9`) and `getNeedleOptions()` sets only auth and proxy
(`RokuDevice.js:258`), so needle's defaults stand — `open_timeout` 10 s, but
`response_timeout` and `read_timeout` both 0, i.e. DISABLED. A device that
accepts the socket and never answers therefore hangs indefinitely. `retryCount`
bounds the number of attempts, not the duration of one; the comment that implied
otherwise is corrected.

`withTimeout` moves to its own leaf module because `diagnostics.js` cannot
import `driver.js`: `driver.js` imports `steps.js` for `sleep` and `steps.js`
imports `diagnostics.js` for `diagnosedError`. Same mechanical reason
`home-list.js` exists. `driver.js` re-exports it, so `registry.js` and
`scripts/measure-signin.js` keep their import site unchanged.

Full RTA suite green on `.177` with this applied: 57 passed / 5 skipped,
`failures: []`, fixture 200 at both ends, `peakLive` 9 inside the documented
9-11 band.
